Cleanup + comments
