+++ frontend/src/components/global.d.ts
@@ -1,3 +1,8 @@
+/**
+ * \file : global.d.ts
+ * \brief : Contains custom type informations for import
+ */
+
 declare module '*.png'{
     const value: import('react-native').ImageURISource;
     export default value;
