+++ frontend/src/components/viewport.tsx
@@ -1,4 +1,5 @@
 /**
+ * \file : viewport.tsx
  * \brief : React module that shows the simulation. 
  */
 import React from "react";
